HW15/requirements.txt
django==4.2.8
Pillow==10.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.0